requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
iso8601==1.0.2
keyboard==0.13.5
mouse==0.7.1
multidict==6.0.2
python-dotenv==0.20.0
twitchio==2.2.0
typing_extensions==4.2.0
yarl==1.7.2
